nvmestore: clean
